Split up logic to run with 'python'
