+++ flask_scratch/routes.py
@@ -1,0 +1,7 @@
+from flask_scratch import app
+
+
+@app.route("/hello")
+def hello():
+    name = request.args.get("name", "world")
+    return f"Hello, {escape(name)}!"
