+++ tests/test_hello.py
@@ -1,27 +1,27 @@
 import pytest
 
-from flask_scratch import create_app
+from flask_scratch.app import create_app
 from http import HTTPStatus
 
 
 @pytest.fixture
-def app():
+def app_in_test():
     app = create_app({"TESTING": True})
     yield app
 
 
 @pytest.fixture
-def client(app):
-    return app.test_client()
+def client_in_test(app_in_test):
+    return app_in_test.test_client()
 
 
-def test_root(client):
-    res = client.get("/")
+def test_root(client_in_test):
+    res = client_in_test.get("/")
     assert res.status_code == HTTPStatus.NOT_FOUND
     return True
 
 
-def test_hello(client):
-    res = client.get("/hello")
+def test_hello(client_in_test):
+    res = client_in_test.get("/hello")
     assert res.status_code == HTTPStatus.OK
     return True
